Move assets directory
